moving DataTable and EntityRow to a subdir
